requirements.txt
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
pynput>=1.7.6
pyperclip>=1.8.2
requests>=2.31.0
noisereduce>=3.0.0
psutil>=5.9.0
pywhispercpp>=1.0.0
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
